--- src/components/Carouselv2/styles.scss
@@ -1,5 +1,0 @@
-.my-swiper {
-  .swiper-slide {
-      max-width: 255px;
-  }
-}
