--- src/components/Carouselv3/index.jsx
@@ -1,58 +1,0 @@
-import React, { useEffect } from 'react'
-import { useLocation, useParams } from 'react-router-dom'
-import { Swiper, SwiperSlide } from 'swiper/react'
-import { Keyboard } from 'swiper'
-import Box from '@mui/material/Box'
-
-import MediaCard from '../MediaCard'
-import useItemsv3 from '../../hooks/useItemsv3'
-import usePath from '../../hooks/usePath'
-
-import 'swiper/css'
-import 'swiper/css/pagination'
-//import 'swiper/css/navigation'
-import './styles.scss'
-
-const Carouselv3 = ({ title }) => {
-  const { show = 'promos' } = useParams()
-  let { pathname } = useLocation()
-  const { getPath } = usePath()
-  const { carouselItems, category, getItemsToShow } = useItemsv3()
-
-  pathname = getPath(pathname)
-
-  useEffect(() => {
-    getItemsToShow(show, pathname)
-  }, [getItemsToShow, show, pathname])
-
-  const carouselClasses = `carousel carousel-${show}`
-  const carouselTitle = title && <h2 className="title-carousel">{title}</h2>
-
-  return (
-    <Box className={carouselClasses} component="section">
-      {carouselTitle}
-
-      <Swiper
-        className="my-swiper"
-        keyboard={{
-          enabled: true,
-        }}
-        modules={[Keyboard]}
-        pagination={{
-          clickable: true,
-        }}
-        slidesPerView={'auto'}
-        spaceBetween={24}
-      >
-        {carouselItems.length > 0 &&
-          carouselItems.map((item, index) => (
-            <SwiperSlide key={index}>
-              <MediaCard {...item} categoria={category.nombre} tipo={category.tipo} />
-            </SwiperSlide>
-          ))}
-      </Swiper>
-    </Box>
-  )
-}
-
-export default Carouselv3
