+++ src/hooks/useItems.js
@@ -2,53 +2,79 @@ import { useContext, useState } from 'react'
 
 import ProductsContext from '../context/ProductsContext'
 
-const useItems = () => {
-  const { products, setProducts, categories, setCategories } = useContext(ProductsContext)
-  const [cosmicArray, setCosmicArray] = useState([])
+const useItemsv3 = () => {
+  const { products, categories, combos, dataLoaded } = useContext(ProductsContext)
+  const [carouselItems, setCarouselItems] = useState([])
+  const [item, setItem] = useState(null)
+  const [category, setCategory] = useState(null)
 
-  const find = (array, key, value) => array.find((element) => element[key] === value)
+  const [prevShow, setPrevShow] = useState(null)
+  const [prevPathName, setPrevPathName] = useState('')
+
   const filter = (array, key, value) => array.filter((element) => element[key] === value)
+  const find = (array, key, value) => array.find((element) => element[key] === value)
+  const some = (array, value) => array.some((element) => element === value)
 
-  const setCosmicArrayToUse = (arrayName) => {
-    switch (arrayName) {
-      case 'combos':
-        // setCosmicArray(combos)   Si el useState es de tipo array, solo debe recibir array
-        break
-      case 'products':
-        setCosmicArray(products)
-        break
+  const getCosmicArrayToUse = (pathName) => {
+    const pathsThatUseProducts = ['/categoria', '/producto', '/veganizado/producto']
+    const pathsThatUseCombos = ['/', '/combos', '/combo', '/veganizado/combo']
+
+    let useCombosArray = some(pathsThatUseCombos, pathName)
+    let useProductsArray = some(pathsThatUseProducts, pathName)
+
+    return { useCombosArray, useProductsArray }
+  }
+
+  const getCategoryToShow = (showName) => {
+    if (dataLoaded) {
+      let categoryToShow = find(categories, 'nombre', showName)
+
+      setCategory(categoryToShow)
     }
   }
 
-  const getItemsToShow = (showName) => {
-    const category = find(categories, 'nombre', showName)
-    const itemsToShow = filter(cosmicArray, 'categoria', category.id)
+  const getItemsToShow = (showName, pathName) => {
+    let { useCombosArray, useProductsArray } = getCosmicArrayToUse(pathName)
 
-    const categoryName = category.nombre
+    if (dataLoaded && (prevPathName !== pathName || prevShow != showName)) {
+      setPrevPathName(pathName)
+      setPrevShow(showName)
 
-    return { itemsToShow, categoryName }
+      let categoryToShow = find(categories, 'nombre', showName)
+      let itemsToShow = null
+
+      useProductsArray && (itemsToShow = filter(products, 'categoria', categoryToShow.id))
+      useCombosArray && (itemsToShow = filter(combos, 'categoria', categoryToShow.id))
+
+      setCategory(categoryToShow)
+      setCarouselItems(itemsToShow)
+    }
   }
 
-  const getItemToShow = (showID) => {
-    let itemToShow = filter(cosmicArray, 'id', showID)
-    const categoria = find(categories, 'id', itemToShow.categoria)
+  const getItemToShow = (showID, pathName) => {
+    let { useCombosArray, useProductsArray } = getCosmicArrayToUse(pathName)
 
-    itemToShow.categoria = categoria.nombre
+    showID = Number(showID)
 
-    return itemToShow
+    if (dataLoaded && (prevPathName !== pathName || prevShow != showID)) {
+      setPrevPathName(pathName)
+      setPrevShow(showID)
+
+      let itemToShow = null
+
+      useProductsArray && (itemToShow = find(products, 'id', showID))
+      useCombosArray && (itemToShow = find(combos, 'id', showID))
+
+      let categoryToShow = find(categories, 'id', itemToShow.categoria)
+
+      itemToShow.categoria = categoryToShow.nombre
+      itemToShow.tipo = categoryToShow.tipo
+
+      setItem(itemToShow)
+    }
   }
 
-  return {
-    products,
-    setProducts,
-    categories,
-    setCategories,
-    cosmicArray,
-    setCosmicArray,
-    setCosmicArrayToUse,
-    getItemToShow,
-    getItemsToShow,
-  }
+  return { carouselItems, category, item, getCategoryToShow, getItemToShow, getItemsToShow }
 }
 
-export default useItems
+export default useItemsv3
