+++ src/pages/CategoryDetails/index.jsx
@@ -2,9 +2,8 @@ import React from 'react'
 
 import Navbar from '../../components/Navbar'
 import CategoryHeader from '../../components/CategoryHeader'
-/* import Carousel from '../../components/Carousel' */
+import Carousel from '../../components/Carousel'
 import Footer from '../../components/Footer'
-import Carouselv3 from '../../components/Carouselv3'
 import '../../App.scss'
 
 const CategoryDetails = () => {
@@ -12,8 +11,7 @@ const CategoryDetails = () => {
     <div className="body">
       <Navbar />
       <CategoryHeader />
-      {/* <Carousel /> */}
-      <Carouselv3 />
+      <Carousel />
       <Footer />
     </div>
   )
