+++ src/pages/Home/index.jsx
@@ -4,12 +4,11 @@ import AboutUs from '../../components/AboutUs'
 import Navbar from '../../components/Navbar'
 import Banner from '../../components/Banner'
 import Founders from '../../components/Founders'
-/* import Carousel from '../../components/Carousel' */
+import Carousel from '../../components/Carousel'
 import VeganInfo from '../../components/VeganInfo'
 import Location from '../../components/Location'
 import ContactUs from '../../components/ContactUs'
 import Footer from '../../components/Footer'
-import Carouselv3 from '../../components/Carouselv3'
 import '../../App.scss'
 
 const Home = () => {
@@ -19,8 +18,7 @@ const Home = () => {
       <Banner />
       <Founders />
       <VeganInfo />
-      {/* <Carousel /> */}
-      <Carouselv3 title="promos" />
+      <Carousel title="promos" />
       <AboutUs />
       <Location />
       <ContactUs />
